requirements.txt
omegaconf
av
matplotlib
pillow
mss
accelerate